Added handlers for /, /apps/public, and /apps/recently-ran
